+++ packages/fedify/src/utils/docloader.ts
@@ -3,7 +3,6 @@ import {
   type DocumentLoader,
   type DocumentLoaderFactoryOptions,
   type DocumentLoaderOptions,
-  getDocumentLoader,
   getRemoteDocument,
   logRequest,
   type RemoteDocument,
@@ -90,50 +89,4 @@ export function getAuthenticatedDocumentLoader(
     return getRemoteDocument(url, response, load);
   }
   return load;
-}
-
-const _fetchDocumentLoader = getDocumentLoader();
-const _fetchDocumentLoader_allowPrivateAddress = getDocumentLoader({
-  allowPrivateAddress: true,
-});
-
-/**
- * A JSON-LD document loader that utilizes the browser's `fetch` API.
- *
- * This loader preloads the below frequently used contexts:
- *
- * - <https://www.w3.org/ns/activitystreams>
- * - <https://w3id.org/security/v1>
- * - <https://w3id.org/security/data-integrity/v1>
- * - <https://www.w3.org/ns/did/v1>
- * - <https://w3id.org/security/multikey/v1>
- * - <https://purl.archive.org/socialweb/webfinger>
- * - <http://schema.org/>
- * @param url The URL of the document to load.
- * @param allowPrivateAddress Whether to allow fetching private network
- *                            addresses.  Turned off by default.
- * @returns The remote document.
- * @deprecated Use {@link getDocumentLoader} instead.
- */
-export function fetchDocumentLoader(
-  url: string,
-  allowPrivateAddress?: boolean,
-): Promise<RemoteDocument>;
-export function fetchDocumentLoader(
-  url: string,
-  options?: DocumentLoaderOptions,
-): Promise<RemoteDocument>;
-export function fetchDocumentLoader(
-  url: string,
-  arg: boolean | DocumentLoaderOptions = false,
-): Promise<RemoteDocument> {
-  const allowPrivateAddress = typeof arg === "boolean" ? arg : false;
-  logger.warn(
-    "fetchDocumentLoader() function is deprecated.  " +
-      "Use getDocumentLoader() function instead.",
-  );
-  const loader = allowPrivateAddress
-    ? _fetchDocumentLoader_allowPrivateAddress
-    : _fetchDocumentLoader;
-  return loader(url);
 }